requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
Faker==26.0.0
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
